// src/features/authorization/domain/errors/authorization-error-code.enum.ts
export enum AuthorizationErrorCode {
  OWNER_IMMUTABLE = 'OWNER_IMMUTABLE',
  OWNER_ROLE_NOT_ASSIGNABLE = 'OWNER_ROLE_NOT_ASSIGNABLE',
  SELF_MANAGEMENT_FORBIDDEN = 'SELF_MANAGEMENT_FORBIDDEN',
  PERMISSION_NOT_HELD = 'PERMISSION_NOT_HELD',
  PERMISSION_RESERVED_TO_OWNER = 'PERMISSION_RESERVED_TO_OWNER',
  INVITATION_NOT_FOUND = 'INVITATION_NOT_FOUND',
  INVITATION_NOT_PENDING = 'INVITATION_NOT_PENDING',
  INVITATION_EXPIRED = 'INVITATION_EXPIRED'
}